streamlit/requirements.txt
streamlit
tensorflow
pillow
numpy
transformers
torch